chore(dev): automatically add release milestone to each PR
Based on the target branch we will determine the target release/milestone.

If the base branch is [0-9].x (dev branch) we set the milestone as the next minor
release for that release line.

If the base branch is [0-9].[0-9]+ (release branch) we set the milestone as the
next patch release for that version.

If the branch does not match one of those patterns, we do not set a milestone.

If a milestone is already set, we do not try to compute and set.